Factor out fakeTopo from vt/vtgate to vt/topo/test/faketopo
